+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Functionality for renaming a file (#193)
 - Styling for a secondary button (#193)
 - The web component `runCompleted` event now returns `duration: null` if the host page's tab loses focus during the code run (#192)
+- Global nav on editor site with link to `raspberrypi.org` and account menu
 
 ### Changed